fix some assertions on 64-bit osx builds
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@953 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueMemoryPool.cpp
@@ -307,7 +307,7 @@ void GOrgueMemoryPool::InitPool()
 	m_PoolSize = 0;
 	CalculatePageSize();
 	CalculatePoolLimit();
-	wxLogDebug(wxT("Memory pool limit: %ld bytes (page size: %d)"), m_PoolLimit, m_PageSize);
+	wxLogDebug(wxT("Memory pool limit: %ld bytes (page size: %d)"), (long int)m_PoolLimit, (int)m_PageSize);
 
 	while (m_PoolLimit)
 	{
